+++ src/OptionT.js
@@ -24,6 +24,12 @@
 
 'use strict';
 
+var dumpWarn = null;
+if (process.env.NODE_ENV !== 'production') {
+    dumpWarn = (typeof console.warn === 'function') ?
+               console.warn : console.log;
+}
+
 /**
  *  @constructor
  *  @template   T
@@ -151,6 +157,8 @@ OptionT.prototype = Object.freeze({
    /**
      *  Applies a function `fn` to the contained value or returns a default `def`.
      *
+     *  XXX: this API is unstable. See: https://github.com/saneyuki/option-t.js/pull/50
+     *
      *  @template   T, U
      *
      *  @param  {U} def
@@ -158,6 +166,10 @@ OptionT.prototype = Object.freeze({
      *  @return {U}
      */
     mapOr: function OptionTMapOr(def, fn) {
+        if (process.env.NODE_ENV !== 'production') {
+            dumpWarn('Option<T>.mapOr() is experimental. This might be breaking changed whenever.');
+        }
+
         if (this.is_some) {
             return fn(this.value);
         }
@@ -169,6 +181,8 @@ OptionT.prototype = Object.freeze({
    /**
      *  Applies a function `fn` to the contained value or computes a default result by `defFn`.
      *
+     *  XXX: this API is unstable. See: https://github.com/saneyuki/option-t.js/pull/50
+     *
      *  @template   T, U
      *
      *  @param  {function():U}  defFn
@@ -176,6 +190,10 @@ OptionT.prototype = Object.freeze({
      *  @return {U}
      */
     mapOrElse: function OptionTMapOrElse(defFn, fn) {
+        if (process.env.NODE_ENV !== 'production') {
+            dumpWarn('Option<T>.mapOrElse() is experimental. This might be breaking changed whenever.');
+        }
+
         if (this.is_some) {
             return fn(this.value);
         }